# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Polo Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments/{tournamentId}/out-dates:
    get:
      operationId: listTournamentOutDates
      tags: [teamOutDates]
      summary: List all out dates for tournament
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Out dates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamOutDate"
  /tournaments/{tournamentId}/teams/{teamId}/out-dates:
    get:
      operationId: listTeamOutDates
      tags: [teamOutDates]
      summary: List out dates for team in tournament
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: teamId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Team out dates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamOutDate"
    post:
      operationId: createTeamOutDate
      tags: [teamOutDates]
      summary: Create team out date
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: teamId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamOutDateRequest"
      responses:
        "201":
          description: Out date created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamOutDate"
  /out-dates/{outDateId}:
    delete:
      operationId: deleteTeamOutDate
      tags: [teamOutDates]
      summary: Delete out date
      parameters:
        - name: outDateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tournaments/{tournamentId}/play-dates:
    get:
      operationId: listPlayDates
      tags: [playDates]
      summary: List play dates
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Play dates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayDate"
    post:
      operationId: createPlayDate
      tags: [playDates]
      summary: Create play date
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayDateRequest"
      responses:
        "201":
          description: Play date created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayDate"
  /play-dates/{playDateId}:
    put:
      operationId: updatePlayDate
      tags: [playDates]
      summary: Update play date
      parameters:
        - name: playDateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayDateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayDate"
    delete:
      operationId: deletePlayDate
      tags: [playDates]
      summary: Delete play date
      parameters:
        - name: playDateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /tournaments/{tournamentId}/generate-schedule:
    post:
      operationId: generateSchedule
      tags: [tournaments]
      summary: Generate tournament schedule
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Schedule generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateScheduleResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Field:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        number:
          type: integer
          nullable: true
        lat:
          type: string
          nullable: true
        lng:
          type: string
          nullable: true
        surfaceType:
          type: string
          nullable: true
        isActive:
          type: boolean
      required:
        - id
        - clubId
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        name:
          type: string
        shortName:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        logoThumbUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        handicap:
          type: string
          nullable: true
        contactName:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - id
        - clubId
        - name
...
    TeamOutDate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        tournamentId:
          type: string
          format: uuid
        outDate:
          type: string
        reason:
          type: string
          nullable: true
        createdByUserId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - teamId
        - tournamentId
        - outDate
    CreateTeamOutDateRequest:
      type: object
      properties:
        outDate:
          type: string
        reason:
          type: string
      required:
        - outDate
    PlayDate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tournamentId:
          type: string
          format: uuid
        date:
          type: string
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        fieldIds:
          type: array
          items:
            type: string
          nullable: true
        lunchStart:
          type: string
          nullable: true
        lunchEnd:
          type: string
          nullable: true
      required:
        - id
        - tournamentId
        - date
    CreatePlayDateRequest:
      type: object
      properties:
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        fieldIds:
          type: array
          items:
            type: string
        lunchStart:
          type: string
        lunchEnd:
          type: string
      required:
        - date
    UpdatePlayDateRequest:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        fieldIds:
          type: array
          items:
            type: string
        lunchStart:
          type: string
        lunchEnd:
          type: string
    MatchWithTeams:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tournamentId:
          type: string
          format: uuid
        homeTeamId:
          type: string
          format: uuid
          nullable: true
        awayTeamId:
          type: string
          format: uuid
          nullable: true
        fieldId:
          type: string
          format: uuid
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        homeScore:
          type: integer
        awayScore:
          type: integer
        currentChukker:
          type: integer
        clockStartedAt:
          type: string
          format: date-time
          nullable: true
        clockElapsedSeconds:
          type: integer
        clockIsRunning:
          type: boolean
        status:
          type: string
          enum: [scheduled, live, halftime, final, postponed, cancelled]
        round:
          type: string
          nullable: true
        bracketPosition:
          type: integer
          nullable: true
        isLocked:
          type: boolean
        notes:
          type: string
          nullable: true
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
        field:
          $ref: "#/components/schemas/Field"
        tournament:
          $ref: "#/components/schemas/TournamentBrief"
      required:
        - id
        - tournamentId
        - homeScore
        - awayScore
        - currentChukker
        - clockElapsedSeconds
        - clockIsRunning
        - status
        - isLocked
    TournamentBrief:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        chukkersPerMatch:
          type: integer
        clubId:
          type: string
          format: uuid
        clubName:
          type: string
      required:
        - id
        - name
        - chukkersPerMatch
        - clubId
        - clubName
...
    GenerateScheduleResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/MatchWithTeams"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleWarning"
      required:
        - matches
        - warnings
    ScheduleWarning:
      type: object
      properties:
        type:
          type: string
          enum: [error, warning]
        message:
          type: string
      required:
        - type
        - message
...